refactor: get access token
