+++ src/App.tsx
@@ -1,56 +1,30 @@
 import { useEffect, useState } from 'react';
-import type { Product } from './api/services/product-service';
-import { productService } from './api/services/product-service';
+import { authService } from './api/services/auth-service';
 
-function App() {
-  const [data, setData] = useState<Product[]>();
+export default function App() {
+  const [token, setToken] = useState<string>();
+  const [authError, setAuthError] = useState<string>();
   const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string>();
 
   useEffect(() => {
-    const fetchProducts = async (): Promise<void> => {
-      const productsData = await productService.getProducts();
-      console.log(productsData);
-
-      const mappedProducts = productsData.results.map(product => {
-        return {
-          id: product.id,
-          key: product.key,
-        };
-      });
-      setData(mappedProducts);
-      setLoading(false);
-    };
-
-    fetchProducts().catch(() => {
-      setError('Error fetching products');
-    });
+    authService
+      .getAccessToken()
+      .then(token => setToken(token.access_token))
+      .catch(() => setAuthError('Не удалось получить токен'))
+      .finally(() => setLoading(false));
   }, []);
 
   return (
     <>
       <h1>eCommerce Application</h1>
-      {loading && <p>Loading...</p>}
-      {error && <p>{error}</p>}
-      {data && (
-        <div>
-          {data.map(element => (
-            <div
-              key={element.id}
-              style={{
-                border: '1px solid black',
-                margin: '10px',
-                padding: '10px',
-              }}
-            >
-              <h3>{element.key}</h3>
-              <p>{element.id}</p>
-            </div>
-          ))}
-        </div>
+
+      {loading && <p>Получаем токен…</p>}
+      {authError && <p style={{ color: 'red' }}>{authError}</p>}
+      {token && (
+        <p>
+          <strong>Token:</strong> {token}
+        </p>
       )}
     </>
   );
 }
-
-export default App;
