+++ src/sources/token.ts
@@ -1,0 +1,6 @@
+export type Token = {
+  access_token: string;
+  token_type: string;
+  expires_in: number;
+  scope: string;
+};
